refactor(report): move reporting functionality to its own package
