+++ application/src/main/java/fi/hsl/parkandride/config/CoreConfiguration.java
@@ -11,6 +11,9 @@ import fi.hsl.parkandride.core.back.*;
 import fi.hsl.parkandride.core.domain.prediction.AverageOfPreviousWeeksPredictor;
 import fi.hsl.parkandride.core.domain.prediction.Predictor;
 import fi.hsl.parkandride.core.service.*;
+import fi.hsl.parkandride.core.service.reporting.FacilityUsageReportService;
+import fi.hsl.parkandride.core.service.reporting.HubsAndFacilitiesReportService;
+import fi.hsl.parkandride.core.service.reporting.MaxUtilizationReportService;
 import org.jasypt.util.password.PasswordEncryptor;
 import org.jasypt.util.password.StrongPasswordEncryptor;
 import org.joda.time.format.ISOPeriodFormat;
