+++ application/src/main/java/fi/hsl/parkandride/core/service/reporting/HubsAndFacilitiesReportService.java
@@ -1,11 +1,12 @@
 // Copyright © 2015 HSL <https://www.hsl.fi>
 // This program is dual-licensed under the EUPL v1.2 and AGPLv3 licenses.
 
-package fi.hsl.parkandride.core.service;
+package fi.hsl.parkandride.core.service.reporting;
 
 import fi.hsl.parkandride.back.RegionRepository;
 import fi.hsl.parkandride.core.back.UtilizationRepository;
 import fi.hsl.parkandride.core.domain.*;
+import fi.hsl.parkandride.core.service.*;
 import fi.hsl.parkandride.front.ReportParameters;
 
 import java.util.ArrayList;
