+++ application/src/main/java/fi/hsl/parkandride/core/service/reporting/MaxUtilizationReportService.java
@@ -1,13 +1,13 @@
 // Copyright © 2015 HSL <https://www.hsl.fi>
 // This program is dual-licensed under the EUPL v1.2 and AGPLv3 licenses.
 
-package fi.hsl.parkandride.core.service;
+package fi.hsl.parkandride.core.service.reporting;
 
 import com.mysema.commons.lang.CloseableIterator;
 import fi.hsl.parkandride.back.RegionRepository;
 import fi.hsl.parkandride.core.back.UtilizationRepository;
 import fi.hsl.parkandride.core.domain.*;
-import fi.hsl.parkandride.core.service.Excel.TableColumn;
+import fi.hsl.parkandride.core.service.*;
 import fi.hsl.parkandride.front.ReportParameters;
 
 import java.util.ArrayList;
@@ -85,7 +85,7 @@ public class MaxUtilizationReportService extends AbstractReportService {
         sort(rows);
 
         Excel excel = new Excel();
-        List<TableColumn<MaxUtilizationReportRow>> columns = asList(
+        List<Excel.TableColumn<MaxUtilizationReportRow>> columns = asList(
                 tcol("reports.utilization.col.hub", (MaxUtilizationReportRow r) -> r.hub.name),
                 tcol("reports.utilization.col.region", (MaxUtilizationReportRow r) -> ctx.regionByHubId.getOrDefault(r.key.targetId, UNKNOWN_REGION).name),
                 tcol("reports.utilization.col.operator", (MaxUtilizationReportRow r) -> r.operatorNames),
