+++ application/src/main/java/fi/hsl/parkandride/core/service/reporting/ReportService.java
@@ -1,7 +1,7 @@
 // Copyright © 2015 HSL <https://www.hsl.fi>
 // This program is dual-licensed under the EUPL v1.2 and AGPLv3 licenses.
 
-package fi.hsl.parkandride.core.service;
+package fi.hsl.parkandride.core.service.reporting;
 
 import fi.hsl.parkandride.core.domain.User;
 import fi.hsl.parkandride.front.ReportParameters;
